[Gluten-1234] Fix error 'Invalid number of columns in chunk pushed to OutputPort' when executing hash agg after union all
Fix error 'Invalid number of columns in chunk pushed to OutputPort' when executing hash agg after union all:
When executing hash agg after union all, the input attributes of the hash agg will be set as aggregateResultAttributes but not the output of the union all operator.

For example: tpcds Q66, Q76.

Close #1234.