[CORE] Refactor fallback policy (#2763)
This pr refactor the fallback policy to resovle several issues:

- A columnar BHJ fallback to row-based BHJ but it has a coumnar broadcast exchange
- A vanilla ColumnarToRow on the top of columnar shuffle exchange
- The fallback number does not count correctly
- The fallback reason does not been catched


The main change is in ColumnarOverrideRules, it splits postColumnarTransitions into 3 parts:

- fallbackPolicy
Used to check if we should fallback according to the config wholeStageFallbackThreshold or queryFallbackThreshold. Make a new rule PlanFallbackPolicy to hold related code
- postOverrides
Same with previous postOverrides
- finallyRules
Used to report fallback reason and cleanup tags. These rules are moved from preOverrides.
This pr also removes the config spark.gluten.sql.columnar.fallback.policy, it is really confused to user. We can decide to use which config by whether we are inside AQE or not.