[VL] Realloc shuffle split buffer if beyond avg threshold (#2982)
To boost performance, we calculate the average binary column's size per row and use the size + fixed column width to estimate the row count each reducer can allocate. Without the PR, we only calculate the binary column's size using the first batch. The problem is the binary size may varies a lot in batches, the first batch's size may be too small or too large. If it's too small, we need to reallocate memory during split which causes OOM.

The PR calculates the average batch size of each input record batch and compare with current allocated buffer size. If it's out of a threshold, we will firstly try to shrink or re-allocate the buffer based on new size, if the count is already larger than new size, then we spill the current record batch and reallocate a new buffer using the average batch size.