[GLUTEN-2919][VL]Fix incorrect hive scan fallback or offload for velox (#2922)
* [GLUTEN-2919][VL]Support orc format in hive scan

Due to some code change for HiveTableScanExecTransformer, scan offloading behavior becomes incorrect, such as ORC format scan can't be offloaded anymore. The root cause is that some ClickHouse format/datatype check logic is added in this Op which is supposed to add in CKBackend API instead. This PR intends to fix it.