[GLUTEN-3328][VL]Add mutex for ResourceMap operators (#3332)
Velox Parquet write operation now utilizes asynchronous writing for improved performance. The batch will be placed into the VeloxWriteQueue in spark task. And a new thread will be spawned to retrieve the batch data, write it to the file, and close the batch after writing. Hence, we need to incorporate a locking mechanism in ResourceMap.

this patch also removed the twice close batch logic.