[GLUTEN-3361] Support spark 3.4 in Gluten (#3360)
Mainly changes:

In Spark 3.4, there was an API change in BatchScanExec. You can find the details here. A new table parameter has been added, so it needs to be included in the shim layer.
We have copied some code from vanilla Spark in FileSourceScanExecTransformer. You can find it here. In Spark 3.3 and 3.2, the copied code was private, but in Spark 3.4, it is protected. Therefore, it should be placed in the shim layer. Additionally, GlutenTimeMetric.scala and Arm.scala need to be placed in the shim layer to resolve compilation issues.
Offset.scala is newly introduced in Spark 3.4. Hence, it should be added to the Spark 3.2 and Spark 3.3 shim layers.
Empty2Null is in org.apache.spark.sql.execution.datasources.FileFormatWriter package in spark 3.2 and spark 3.3. So we also need add Empty2Null in org.apache.spark.sql.execution.datasources.FileFormatWriter package in Spark 3.4 to pass the compile issue.
The PartitionedFile API changed the filePath parameter from string to SparkPath object in 3.4.
In Spark 3.4, StatFunctions.scala is no longer needed in shim layer. Therefore, it has been moved from shim common to spark 3.3 and shim spark 3.2 shim layer.
PromotePrecision.scala has been deleted in Spark 3.4.
The unused DwrfScan code has been removed from the Spark 3.2 and Spark 3.3 shim layers.
Since DataSourceStrategyUtil.scala is not used in gluten, it has been deleted.
A new API, visitOffset(p: Offset), has been added to the LogicalPlanVisitor in Spark 3.4.