[GLUTEN-3598][CH] Support to config the hash algorithm for the ch shuffle hash partitioner
Now the hash algorithm of the ch shuffle hash partitioner is cityHash64, which is different from vanilla spark, when there is one side shuffle of the join fallbacking, the hash id are different between the ch and vanilla spark, so add a configuration to control the hash algorithm for the ch shuffle hash partitioner.

Close #3598.