[GLUTEN-3779][CH] Fix core dump when executing sql with runtime filter
When there are more than one input row in the final stage with the bloom_filter_agg, it will be core dump for the CH backend.

The RC is: when merging values in the final stage, the input data maybe a non-init AggregateFunctionGroupBloomFilterData, it will use the wrong filter size and filter_hashes values to init the first AggregateFunctionGroupBloomFilterData, which leads to set the wrong filter size when merging values.

Close #3779.