[VL] Fix incorrect url path caused by decoding on non-original encoded path in PartitionedFile (#3873)
When constructing Spark PartitionedFile, the path used is the not the original encoded path so the decoded path in IteratorApiImpl is not correct in some cases. Current we observed the case for '+'. It's encoded to %2B in URLEncoder.encode but it becomes + when passed into PartitionedFile and finally becomes   after decode. This PR tries to fix this issue with UT added.