[GLUTEN] Remove FIXME comment in `SparkResourceUtil#getTaskSlots` function (#4439)
The SparkResourceUtil#getTaskSlots function in the codebase has a FIXME comment, expressing concern about its performance when dynamic allocation is enabled. After further analysis and understanding of Spark's dynamic allocation implementation, it has been determined that the number of executors may be changed, but the cores of each executor remains constant with dynamic allocation.