[CORE] Remove tag TRANSFORM_SUPPORTED and rename transform-hint to fallback-hint: Phase 1 (#4731)
Since we now introduced a set of validation rules and backend-specific validation rules that all tend to add transform-hint tags to plan nodes, it more or less becomes confusing to tag a plan as TRANSFORM_SUPPORTED. Before the plan is finally implemented within Gluten transformers, a plan node can be turned from TRANSFORM_SUPPORTED to TRANSFORM_UNSUPPORTED in any particular validation rule. So if a plan is tagged as TRANSFORM_SUPPORTED, it's never guaranteed to be executed natively by Gluten.

To eliminate this confusing state and ease understanding, the PR (and next PR as phase 2) targets to remove tag TRANSFORM_SUPPORTED and rename the term transform-hint to fallback-hint. When a plan is tagged with the hint, then it's considered a plan to be fallback and being run in Vanilla Spark.

Phase 1 (this PR):
Remove tag TRANSFORM_SUPPORTED.

Phase 2 (in next PR):
Rename transform-hint to fallback-hint.