[GLUTEN-5024][VL] Enhance buildbundle-veloxbe.sh to have the ability to run only one specific step
buildbundle-veloxbe.sh is composed of at least two steps:

- Get and build Velox
- Build Gluten CPP code

Currently it is not smooth to run on some platforms such as MacOS, we might
need to run this script multiple times to fix a build issue, e.g. `Get and build Velox`
succeed, but `Build Gluten CPP code` failed, in this PR we introduced the ability
to specify the commands to run rather than run the whole script every time.
By default we still run all the steps. Usage samples are the following:

```bash
-- Run all the steps
./dev/buildbundle-veloxbe.sh --build_type=RelWithDebInfo
-- Only run build_gluten_cpp
./dev/buildbundle-veloxbe.sh --build_type=RelWithDebInfo build_gluten_cpp
```