[GLUTEN-5219][CH]Fix the table metadata sync issue for the CH backend
The CH backend receives the table metadata by the substrait plan when executing create table and query, but now there is no version info to represent the table metadata, when it meets the below case, it will execute failed:
```
1. create a table_1 with the path1, insert data
2. query from the table_1
3. drop the table_1 and re-create a table named table_1, but the table path is path2;
4. insert data, will be found that the inserted data will be located in the path1 but not in the path2;
```
So add a timestamp field to represent the version of the table metadata, this timestamp will be get from the Delta snapshot. The CH backend will update the table metadata according to this timestamp.

Close #5219.