[GLUTEN-6253] Use internal udf config to avoid modify the original one (#6255)
The current implementation sets spark.gluten.sql.columnar.backend.velox.udfLibraryPaths on driver side after resolving the library paths. This approach can overwrite the original settings with a local file path on the driver node before sending the SparkConf to all executors, and the executors on different nodes will fail while accessing that path.

This PR sets the resolved library paths to an internal config to avoid the conflicts.

Manually verified on a multi-node cluster.