[CORE] Enlarge defaultRecursionLimit by pre-loading the protobuf class (#8904)
We previously introduced custom protobuf dependency with only defaultRecursionLimit enlarged to avoid the limit is hit for deeply nested plans (one user reported this issue). It is not easy to upgrade this dependency since it requires local build and then publishing the jar to maven repo.

With this change, we can remove the custom protobuf dependency for easily upgrading protobuf when needed.