[GLUTEN-10607][MINOR] Fix: Use setSafe in DateWriter to avoid overflow (#10581)
If the Date type is nested within an Array, Map, or Struct, invoking writeRow may result in an exception due to DateWriter utilizing the unsafe set interface. The set interface does not reallocate the value buffer in cases of overflow. Consequently, this pull request changes to use the safer setSafe interface.

Co-authored-by: 蒋添 <jiangtian@MBP-JXQ2GYYNJT-1934.local>