[GLUTEN-10621][VL] feat: Support cudf parquet connector (#10622)
Cudf connector supports parquet file format, and does not support partition.
The splitInfo partitionColumns should be empty for unpartitioned table, but for iceberg split, it may access the partition columns even if it is unpartitioned table, so still set the partitionColumns field, and check it if map in it is empty to check if it is cudf not supported partition column.
Test on local.