[GLUTEN] Delete temporary files generated by dbgen (#10783)
When using TPCH shell script to generate test data, the script got stuck when execute it again.
The tool dbGen will generate a lots of temporary files and not be deleted after the first running. When running the tool again, it asks whether to overwrite the temporary files, which make the script stuck.

The patch adds a function to delete these temporary files before running the dbGen tool.