[VL] Make VeloxResizeBatchesExec not inherit ColumnarToColumnarTransition
It's an improvement following #10763. The PR decouples ColumnarToColumnarExec from ColumnarToColumnarTransition to make them both can be solely implemented. If a class inherits from ColumnarToColumnarExec but not from ColumnarToColumnarTransition at the same time, it will not be considered as a transition, so it's not going to be removed by RemoveTransitions rule.