[GLUTEN-10933][VL] Support resize GPU batch and shuffle reader outputs cudf::table (#11090)
Now in the post rule, we can only get the stage level plan, so we cannot decide if the ColumnarExchange output should be cudf format, currently, we suppose the ColumnarExchange is always CudfColumnarExchange, will support validation in injectQueryStagePrepRule where we can see the total plan, generated the transformed plan and validate it, then discard it.
After CudfColumnarExchange, concat the batches in GpuResizeBufferColumnarBatchExec to output cudf::table with batch size maximum integer, velox-cudf does not control the batch size now. So the CPU shuffle reader does the decompression and prepare the first batch while GPU executes the CPU tasks one by one.
Performance:
After this change, the stage 2 in TPCDS Q95 SF100 time decreases from 27s to 13s.

In the shuffle writer, split the bool to byte, and timestamp to nanoseconds int64_t to match the cudf format.
Now BroadcastExchange outputs RowVector, so the CudfFromVeclox operator should accept RowVector and CudfVector with this commit, oap-project/velox@3d6ab40

Also fix the Gluten build with GPU.

Related issue: #10933